BE/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
asyncpg==0.30.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bencodepy==0.9.5
bleach==6.0.0
blinker==1.7.0
bs4==0.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==43.0.1
cycler==0.11.0
dacite==1.8.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databases==0.9.0
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.115.5
fastjsonschema==2.17.1
filelock==3.16.1
Flask==3.0.2
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gamsapi==45.3.0
gamspy==0.10.4
gamspy-base==45.3.0
gensim==4.3.3
greenlet==3.0.3
gurobipy==10.0.3
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.2
idna==3.4
ImageHash==4.3.1
importlib_metadata==7.1.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.4
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.114
lazypredict==0.2.12
lightgbm==4.5.0
llvmlite==0.43.0
MarkupSafe==2.1.3
marshmallow==3.22.0
maskpass==0.3.7
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.5
multimethod==1.12
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.16.3
nbformat==5.9.0
neo4j==5.24.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.9.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.60.0
numpy==1.25.0
odfpy==1.4.1
openpyxl==3.1.5
oracledb==2.4.1
orjson==3.10.7
overrides==7.3.1
packaging==24.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.4
phik==0.12.4
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.20.0
prettytable==3.10.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.10
pure-eval==0.2.2
pycoingecko==3.1.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.15.1
pynput==1.7.7
pyparsing==3.1.0
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.7.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
regex==2024.9.11
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.41.2
statsmodels==0.14.2
sympy==1.13.3
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.3.0
tinycss2==1.2.1
torch==2.4.1
torch-geometric==2.6.0
tornado==6.3.2
tqdm==4.66.5
traitlets==5.9.0
typeguard==4.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.32.0
visions==0.7.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.1.1
yarl==1.9.11
ydata-profiling==4.9.0
zipp==3.18.1
